+++ apps/explorer/src/hooks/euler/client.ts
@@ -1,14 +1,14 @@
 import { RPC_URLS, VIEM_CHAINS } from '@cowprotocol/common-const'
 import { FiniteMap } from '@cowprotocol/common-utils'
-import { SupportedChainId } from '@cowprotocol/cow-sdk'
+import { EvmChains } from '@cowprotocol/cow-sdk'
 
 import { createPublicClient, http } from 'viem'
 
 type PublicClient = ReturnType<typeof createPublicClient>
 
-const publicClientsCache = new FiniteMap<SupportedChainId, PublicClient>(16)
+const publicClientsCache = new FiniteMap<EvmChains, PublicClient>(16)
 
-export function getPublicClient(chainId: SupportedChainId): PublicClient {
+export function getPublicClient(chainId: EvmChains): PublicClient {
   const cached = publicClientsCache.get(chainId)
   if (cached) return cached