+++ libs/common-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [3.4.1](https://github.com/cowprotocol/cowswap/compare/common-utils-v3.4.0...common-utils-v3.4.1) (2026-05-22)
+
+### Patch Changes
+
+- fix: patch estimated gas value to hex-encode (#7524)
+
+- fix: detect failed onchain TXs (#7421)
+
+- Updated dependencies []:
+  - @cowprotocol/common-const@3.3.1
+  - @cowprotocol/currency@1.0.1
+  - @cowprotocol/types@4.3.2
+
 ## 3.4.0
 
 ### Minor Changes