+++ libs/wallet/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [3.4.0](https://github.com/cowprotocol/cowswap/compare/wallet-v3.3.2...wallet-v3.4.0) (2026-05-22)
+
+### Minor Changes
+
+- feat: improve safe wallet detection (#7544)
+
+- feat(wallet): add ConnectWalletButton and isConnectionRestoring state (#7468)
+
+### Patch Changes
+
+- fix: patch estimated gas value to hex-encode (#7524)
+
+- fix(wallet): use Eip6963 info from viem (#7538)
+
+- Updated dependencies []:
+  - @cowprotocol/ui@3.6.0
+  - @cowprotocol/common-const@3.3.1
+  - @cowprotocol/common-hooks@3.2.4
+  - @cowprotocol/common-utils@3.4.1
+  - @cowprotocol/core@3.3.0
+  - @cowprotocol/currency@1.0.1
+  - @cowprotocol/ens@3.2.4
+  - @cowprotocol/types@4.3.2
+  - @cowprotocol/iframe-transport@2.2.6
+
 ## 3.3.2
 
 ### Patch Changes